backend/requirements.txt
django-rest-framework
django-cors-headers
django-crontab
speedtest-cli
transformers
python-nmap
matplotlib
colorama
requests
werkzeug
networkx
traffic
pandas
django
torch
scapy
